Validate trigger paths in evaluation script instead of test project
Move path validation into EvaluateConditionalTestScopes.cs which runs in the
pipeline with full repo access. Emits ##[warning] for stale paths. Remove the
Microsoft.NET.Infra.Tests project since Helix payloads don't include the source
tree needed for directory existence checks.
